backend/requirements.txt
flask
flask-cors
flask-limiter
flask-socketio
sqlalchemy
psycopg2-binary
firebase-admin
python-dotenv
pandas
numpy
scikit-learn
xgboost
shap
matplotlib
joblib
celery
redis
networkx

